+++ pytsite/contact/__init__.py
@@ -1,6 +1,6 @@
 """PytSite Contact Form.
 """
-from pytsite import assetman as _assetman, lang as _lang, tpl as _tpl
+# Public API
 from ._form import Contact as ContactForm
 
 __author__ = 'Alexander Shepetko'
@@ -8,8 +8,16 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
-_assetman.register_package(__name__)
-_lang.register_package(__name__)
-_tpl.register_package(__name__)
+def __init():
+    from pytsite import assetman, lang, tpl, browser
 
-_assetman.add('pytsite.contact@js/common.js', forever=True)
+    assetman.register_package(__name__)
+    lang.register_package(__name__)
+    tpl.register_package(__name__)
+
+    assetman.add('pytsite.contact@js/common.js', forever=True)
+
+    browser.register_ep('pytsite.contact.ep.submit')
+
+
+__init()