+++ pytsite/contact/res/assets/js/common.js
@@ -2,7 +2,7 @@ $(function () {
     $('#pytsite-contact-form').on('pytsite_form_submit', function (e, form) {
         form.removeClass('ready-to-submit');
 
-        pytsite.js.post('pytsite.contact.ep.submit', form.serializeForm(), function(response) {
+        pytsite.browser.post('pytsite.contact.ep.submit', form.serializeForm(), function(response) {
             alert(response);
             form[0].reset();
         }, function() {