+++ pytsite/content/_model.py
@@ -676,4 +676,6 @@ class ContentSubscriber(_odm.Model):
         """
         self._define_field(_odm.field.String('email', nonempty=True))
         self._define_field(_odm.field.Bool('enabled', default=True))
-        self._define_index([('email', _odm.I_ASC)])
+        self._define_field(_odm.field.String('language', nonempty=True))
+
+        self._define_index([('email', _odm.I_ASC)], unique=True)