+++ pytsite/content/res/assets/js/content.js
@@ -13,7 +13,7 @@ $(function () {
         var model = $(this).data('model');
         var id = $(this).data('entityId');
         if (model && id) {
-            pytsite.js.post('pytsite.content.ep.view_count', {
+            pytsite.browser.post('pytsite.content.ep.view_count', {
                 model: model,
                 id: id
             });