+++ pytsite/flag/__init__.py
@@ -11,7 +11,7 @@ __license__ = 'MIT'
 
 def __init():
     import sys
-    from pytsite import assetman, odm, tpl, lang, events
+    from pytsite import assetman, odm, tpl, lang, events, browser
     from . import _model, _eh
 
     lang.register_package(__name__)
@@ -26,5 +26,7 @@ def __init():
     assetman.add('pytsite.flag@js/common.js', forever=True)
 
     events.listen('pytsite.odm.entity.delete', _eh.odm_entity_delete)
+
+    browser.register_ep('pytsite.flag.ep.toggle')
 
 __init()